INT: remove complex stream name polling threads
This complex and general code was needed for polling m3u URLs for stream names
in a separate thread. Since we limit to json playlists with stream names, we
can remove all that. Move the fill_stream_names() functionality (setting
stream['name']) to the load_streams() method directly. Rename
poll_stream_names() -> display_stream_names() since that was what it did: wait
for all stream names to be there (now in load_streams()) and then update the
display (Tk listbox).